sandbox/雑談部屋/requirements.txt
streamlit
google-generativeai
openai
anthropic
